+++ src/mass_driver/forges/dummy.py
@@ -30,6 +30,11 @@ class DummyForge(Forge):
         # Return-type of Forge not pinned down
         return DUMMY_PR_URL
 
-    def get_pr(self, forge_repo: str, pr_id: str):
+    def get_pr_status(self, pr_url) -> str:
         """Get the PR by ID on forge_repo"""
-        return DUMMY_PR_URL
+        return "merged"
+
+    @property
+    def pr_statuses(self) -> list[str]:
+        """List the possible PR status returned by get_pr_status, sorted by completion"""
+        return ["merged", "open"]