+++ src/mass_driver/forges/github.py
@@ -1,5 +1,7 @@
 """Githubas Forge. Using the github lib if available"""
 
+import re
+
 from github import AppAuthentication, Github
 from pydantic import SecretStr
 
@@ -32,7 +34,37 @@ class GithubBaseForge(Forge):
         )
         return pr.html_url
 
-    def get_pr(self, forge_repo: str, pr_id: str):
+    def get_pr_status(self, pr_url: str) -> str:
+        """Get the status of a single given PR, used as key to group PRs by status"""
+        owner, repo, pr_num = detect_pr_info(pr_url)
+        owner_repo = f"{owner}/{repo}"
+        pr_obj = self._get_pr(owner_repo, pr_num)
+        if pr_obj.merged:
+            return "merged"
+        if pr_obj.state == "closed":
+            return "closed (but not merged)"
+        if pr_obj.mergeable:
+            return "mergeable (no conflict)"
+        return "non-mergeable (conflicts)"
+
+    @property
+    def pr_statuses(self) -> list[str]:
+        """List possible PR statuses that will be returned by get_pr_status.
+
+        List is sorted from most complete (accepted-and-merged) to least completed (not
+        merged, not review-approved, has merge-conflicts).
+
+        The returned list's ordering is used by the view-pr mass-driver command to show
+        the PRs by status, from most completed to least completed.
+        """
+        return [
+            "merged",
+            "closed (but not merged)",
+            "mergeable (no conflict)",
+            "non-mergeable (conflicts)",
+        ]
+
+    def _get_pr(self, forge_repo: str, pr_id: str):
         """Get the PR by ID on forge_repo"""
         repo = self._github_api.get_repo(forge_repo)
         return repo.get_pull(int(pr_id))
@@ -72,9 +104,6 @@ class GithubAppForge(GithubBaseForge):
         self._github_api = Github(app_auth=auth)
 
 
-# FIXME: Github App login blocked by "no such module JWT.encode"
-
-
 def detect_github_repo(remote_url: str):
     """Find the github remote from a cloneable URL
 
@@ -87,3 +116,19 @@ def detect_github_repo(remote_url: str):
         )
     _junk, gh_name = remote_url.split(":")
     return gh_name.removesuffix(".git")
+
+
+def detect_pr_info(pr_url: str) -> tuple[str, str, str]:
+    """Detect a PR's repo and number
+
+    >>> detect_pr_info("https://github.com/OverkillGuy/sphinx-needs-test/pull/1")
+    ('OverkillGuy', 'sphinx-needs-test', '1')
+    """
+    GITHUB_PR_REGEX = re.compile(
+        r"""https://github.com/([a-zA-Z0-9Z_\.-]+)/([a-zA-Z0-9_\.-]+)/pull/([0-9]+)"""
+    )
+    match = re.fullmatch(GITHUB_PR_REGEX, pr_url)
+    if not match:
+        raise ValueError(f"PR URL {pr_url} doesn't map to github PR regex")
+    owner, repo, prnum = match.groups()
+    return (owner, repo, prnum)